refactor(oxfmt): move Doc-to-IR `postprocess()` under `prettier_compat` (#25141)
Just relocated.
